Python_for_AI/LangChain/First_Project/pyproject.toml
[project]
name = "first-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain[openai]",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[tool.uv.sources]
langchain = { workspace = true }
